refactor: :hammer: removed user module
