+++ src/database/base.py
@@ -1,6 +1,6 @@
 # pylint: disable=unused-import
 from src.database.base_class import Base  # noqa: F401
+from src.modules.auth.models.user import User  # noqa: F401
 from src.modules.auth.models.user_token import UserToken  # noqa: F401
-from src.modules.user.models.user import User  # noqa: F401
 
 # pylint: enable=unused-import
