+++ src/modules/auth/auth_bearers/auth_bearer_with_cookie.py
@@ -1,8 +1,6 @@
-from __future__ import annotations
-
 from typing import Optional
 
-from fastapi import Depends, HTTPException, status
+from fastapi import Depends, HTTPException, Request, status
 from fastapi.openapi.models import OAuthFlows as OAuthFlowsModel
 from fastapi.security import OAuth2
 from sqlalchemy.orm import Session
@@ -25,7 +23,7 @@ class OAuth2PasswordBearerWithCookie(OAuth2):
         super().__init__(flows=flows, scheme_name=scheme_name, auto_error=auto_error)
 
     async def __call__(
-        self, request, db: Session = Depends(get_session)
+        self, request: Request, db: Session = Depends(get_session)
     ) -> Optional[str]:
         access_token: str = request.cookies.get("access_token")
         authorization: bool = verify_jwt_token(db, access_token, "access_token")
@@ -37,5 +35,4 @@ class OAuth2PasswordBearerWithCookie(OAuth2):
                     detail="Not authenticated",
                     headers={"WWW-Authenticate": "Bearer"},
                 )
-
             return None
